refactor: restructure replace-links module and improve namespace resolution
