+++ package.json
@@ -7,7 +7,7 @@
 		"dev": "node esbuild.config.mjs",
 		"build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
 		"version": "node version-bump.mjs && git add manifest.json versions.json",
-		"test": "VITE_CJS_IGNORE_WARNING=true vitest run"
+		"test": "VITE_CJS_IGNORE_WARNING=true vitest"
 	},
 	"keywords": [],
 	"author": "",
