+++ src/replace-links/index.ts
@@ -1,0 +1,2 @@
+export { getEffectiveNamespace, replaceLinks } from "./replace-links";
+export type { ReplaceLinksOptions } from "./types";
