+++ src/test-helpers.ts
@@ -1,0 +1,20 @@
+import { PathAndAliases } from "./path-and-aliases.types";
+import { getEffectiveNamespace } from "./replace-links";
+
+export const getSortedFiles = ({
+	fileNames,
+	restrictNamespace = false,
+	baseDir,
+}: {
+	fileNames: string[];
+	restrictNamespace?: boolean;
+	baseDir?: string;
+}): PathAndAliases[] => {
+	const sortedFileNames = fileNames.slice().sort((a, b) => b.length - a.length);
+	return sortedFileNames.map((path) => ({
+		path,
+		aliases: null,
+		restrictNamespace: restrictNamespace ?? false,
+		namespace: getEffectiveNamespace(path, baseDir),
+	}));
+};
